django_mcp_workshop/requirements/local.txt
-r base.txt
# Development
django-debug-toolbar>=4.3,<5
django-extensions>=3.2,<4
# Code quality
black>=24.0,<25
isort>=5.13,<6
flake8>=7.0,<8
# Testing
pytest-django>=4.8,<5
factory-boy>=3.3,<4
# Pre-commit
pre-commit>=3.7,<4
